backend/requirements.txt
Django>=4.2
djangorestframework
psycopg2-binary
python-dotenv
corsheaders
drf-yasg
dj-database-url
openai
pypdf
gunicorn